requirements.txt
pandas
seaborn
matplotlib
fastapi
sqlalchemy
pytest
uvicorn
sqlalchemy-utils
python-dotenv
httpx
numpy
psycopg2-binary
jupyter
passlib
argon2_cffi
PyJWT
pytest-mock
pydantic
